requirements.txt
kafka-python
scikit-learn
pandas
numpy
joblib
gradio
plotly
minio
dask[complete]
imbalanced-learn